requirements.txt
pyyaml
pyfiglet
requests
gitpython
paramiko
tendo